Add "Completed" badge and greyed out background for a completed task

+++ src/components/todo-item/TodoItem.js
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
+import Badge from 'react-bootstrap/Badge';
 import Button from 'react-bootstrap/Button';
 
 import 'bootstrap-icons/font/bootstrap-icons.css';
@@ -8,13 +9,17 @@ import 'bootstrap-icons/font/bootstrap-icons.css';
 import ConfirmDeleteTask from '../modal/ConfirmDeleteTask';
 
 function TodoItem({ id, title, detail, isComplete }) {
-  const [show, setShow] = useState(false);
+  const [showConfirmModal, setShowConfirmModal] = useState(false);
 
-  const handleShow = () => setShow(true);
-  const handleHide = () => setShow(false);
+  const handleShowConfirmModal = () => setShowConfirmModal(true);
+  const handleHideConfirmModal = () => setShowConfirmModal(false);
 
   const handleDeleteTask = (id, title) => {
     console.log(`Deleting task with id ${id} and title ${title}`);
+  };
+
+  const handleMarkComplete = (id) => {
+    console.log(`Mark task with id ${id} as complete`);
   };
 
   const navigate = useNavigate();
@@ -22,27 +27,41 @@ function TodoItem({ id, title, detail, isComplete }) {
   return (
     <>
       <div className="text-center text-sm-start">
+        {isComplete && <Badge bg="success" className="mb-3">Completed</Badge>}
         <h2 className="mb-2 mb-sm-3 mb-md-4 fw-bold">{title}</h2>
         <p className="fs-5 fst-italic">{detail}</p>
         {/* <p>{isComplete ? 'completed' : 'pending'}</p> */}
       </div>
       <div className="d-flex gap-3">
-        <Button variant="primary" size="lg" onClick={() => navigate(`/edit/${id}`)}>
+        <Button
+          variant="primary"
+          size="lg"
+          onClick={() => navigate(`/edit/${id}`)}
+          disabled={isComplete}
+        >
           <i className="bi bi-pencil-square"></i>
         </Button>
-        <Button variant="danger" size="lg" onClick={handleShow}>
+        <Button
+          variant="danger"
+          size="lg"
+          onClick={handleShowConfirmModal}
+        >
           <i className="bi bi-trash"></i>
         </Button>
-        <Button variant="success" size="lg">
+        <Button
+          variant="success"
+          size="lg"
+          onClick={() => handleMarkComplete(id)}
+        >
           <i className="bi bi-check2-square"></i>
         </Button>
       </div>
-      {show &&
+      {showConfirmModal &&
         <ConfirmDeleteTask
           id={id}
           title={title}
-          show={show}
-          onHide={handleHide}
+          show={showConfirmModal}
+          onHide={handleHideConfirmModal}
           onConfirmDeleteClick={(id, title) => handleDeleteTask(id, title)}
         />
       }
